+++ src/SME.SGP.WebClient/src/dtos/index.js
@@ -3,3 +3,4 @@ export { default as RotasDto } from './rotasDto';
 export { default as periodo } from './periodo';
 export { default as ModalidadeDTO } from './modalidade';
 export { default as tipoEscolaDTO } from './tipoEscolaDto';
+export { default as situacaoAEE } from './situacaoAEE';
